+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fire editable events on the L.Editable instance also
 - added featuresLayer option to L.Editable, and by default add features to
   this layer instead of the map directly
+- added lineGuideOptions
 
 ## 0.3.0
 - added optional latlng param to L.Editable.startPolyline and startPolygon